+++ LibraryApp/Controllers/ReservationsController.cs
@@ -1,7 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -22,41 +18,60 @@ namespace LibraryApp.Controllers
         // GET: Reservations
         public async Task<IActionResult> Index()
         {
-            var libraryDbContext = _context.Reservations.Include(r => r.Book).Include(r => r.Member);
-            return View(await libraryDbContext.ToListAsync());
+            var reservations = await _context.Reservations
+                .Include(r => r.Book)
+                .Include(r => r.Member)
+                .OrderByDescending(r => r.ReservedDate)
+                .ToListAsync();
+            return View(reservations);
         }
 
         // GET: Reservations/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var reservation = await _context.Reservations
-                .Include(r => r.Book)
-                .Include(r => r.Member)
+            if (id == null) return NotFound();
+            var r = await _context.Reservations
+                .Include(r => r.Book).Include(r => r.Member)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (reservation == null)
-            {
-                return NotFound();
-            }
+            if (r == null) return NotFound();
+            return View(r);
+        }
 
-            return View(reservation);
+        // POST: Reservations/Fulfill/5
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Fulfill(int id)
+        {
+            var reservation = await _context.Reservations.FindAsync(id);
+            if (reservation == null) return NotFound();
+            reservation.Status = "Fulfilled";
+            await _context.SaveChangesAsync();
+            TempData["Success"] = "Reservation fulfilled successfully.";
+            return RedirectToAction(nameof(Index));
+        }
+
+        // POST: Reservations/Cancel/5
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Cancel(int id)
+        {
+            var reservation = await _context.Reservations.FindAsync(id);
+            if (reservation == null) return NotFound();
+            reservation.Status = "Cancelled";
+            await _context.SaveChangesAsync();
+            TempData["Success"] = "Reservation cancelled.";
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Reservations/Create
         public IActionResult Create()
         {
-            ViewData["BookId"] = new SelectList(_context.Books, "Id", "Author");
-            ViewData["MemberId"] = new SelectList(_context.Members, "Id", "Email");
+            ViewData["BookId"] = new SelectList(_context.Books, "Id", "Title");
+            ViewData["MemberId"] = new SelectList(_context.Members, "Id", "FullName");
             return View();
         }
 
         // POST: Reservations/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,BookId,MemberId,ReservedDate,ExpiryDate,Status")] Reservation reservation)
@@ -65,86 +80,55 @@ namespace LibraryApp.Controllers
             {
                 _context.Add(reservation);
                 await _context.SaveChangesAsync();
+                TempData["Success"] = "Reservation created.";
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["BookId"] = new SelectList(_context.Books, "Id", "Author", reservation.BookId);
-            ViewData["MemberId"] = new SelectList(_context.Members, "Id", "Email", reservation.MemberId);
+            ViewData["BookId"] = new SelectList(_context.Books, "Id", "Title", reservation.BookId);
+            ViewData["MemberId"] = new SelectList(_context.Members, "Id", "FullName", reservation.MemberId);
             return View(reservation);
         }
 
         // GET: Reservations/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var reservation = await _context.Reservations.FindAsync(id);
-            if (reservation == null)
-            {
-                return NotFound();
-            }
-            ViewData["BookId"] = new SelectList(_context.Books, "Id", "Author", reservation.BookId);
-            ViewData["MemberId"] = new SelectList(_context.Members, "Id", "Email", reservation.MemberId);
+            if (reservation == null) return NotFound();
+            ViewData["BookId"] = new SelectList(_context.Books, "Id", "Title", reservation.BookId);
+            ViewData["MemberId"] = new SelectList(_context.Members, "Id", "FullName", reservation.MemberId);
             return View(reservation);
         }
 
         // POST: Reservations/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,BookId,MemberId,ReservedDate,ExpiryDate,Status")] Reservation reservation)
         {
-            if (id != reservation.Id)
-            {
-                return NotFound();
-            }
-
+            if (id != reservation.Id) return NotFound();
             if (ModelState.IsValid)
             {
-                try
-                {
-                    _context.Update(reservation);
-                    await _context.SaveChangesAsync();
-                }
+                try { _context.Update(reservation); await _context.SaveChangesAsync(); }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!ReservationExists(reservation.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!_context.Reservations.Any(e => e.Id == reservation.Id)) return NotFound();
+                    throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["BookId"] = new SelectList(_context.Books, "Id", "Author", reservation.BookId);
-            ViewData["MemberId"] = new SelectList(_context.Members, "Id", "Email", reservation.MemberId);
+            ViewData["BookId"] = new SelectList(_context.Books, "Id", "Title", reservation.BookId);
+            ViewData["MemberId"] = new SelectList(_context.Members, "Id", "FullName", reservation.MemberId);
             return View(reservation);
         }
 
         // GET: Reservations/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var reservation = await _context.Reservations
-                .Include(r => r.Book)
-                .Include(r => r.Member)
+            if (id == null) return NotFound();
+            var r = await _context.Reservations
+                .Include(r => r.Book).Include(r => r.Member)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (reservation == null)
-            {
-                return NotFound();
-            }
-
-            return View(reservation);
+            if (r == null) return NotFound();
+            return View(r);
         }
 
         // POST: Reservations/Delete/5
@@ -153,18 +137,9 @@ namespace LibraryApp.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var reservation = await _context.Reservations.FindAsync(id);
-            if (reservation != null)
-            {
-                _context.Reservations.Remove(reservation);
-            }
-
+            if (reservation != null) _context.Reservations.Remove(reservation);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool ReservationExists(int id)
-        {
-            return _context.Reservations.Any(e => e.Id == id);
         }
     }
 }
